reorg code structure
